JDBC implementation (SEE EXTENDED DESCRIPTION)
While the code is functional, I'm in the process of refactoring it to query and insert usernames and passwords in a mysql database for the purposes of logging in and registering new users. 

To test these two files, you will need to install connector/j and link it into the project. I use IntelliJ, so I'm not certain how to do this on Eclipse.